+++ kitty_tests/__init__.py
@@ -74,12 +74,15 @@ class BaseTest(TestCase):
     ae = TestCase.assertEqual
     maxDiff = 2000
 
-    def create_screen(self, cols=5, lines=5, scrollback=5, cell_width=10, cell_height=20, options=None):
+    def set_options(self, options=None):
         final_options = {'scrollback_pager_history_size': 1024, 'click_interval': 0.5}
         if options:
             final_options.update(options)
         options = Options(merge_configs(defaults._asdict(), final_options))
         set_options(options)
+
+    def create_screen(self, cols=5, lines=5, scrollback=5, cell_width=10, cell_height=20, options=None):
+        self.set_options(options)
         c = Callbacks()
         return Screen(c, lines, cols, scrollback, cell_width, cell_height, 0, c)